requirements.txt
matplotlib~=3.9.2
python-dotenv~=1.0.1
tinkoff~=0.1.1
pandas~=2.2.3